Move custom eslint rules
