+++ workspaces/eslint-config-rcx-dev/package.json
@@ -17,7 +17,7 @@
     "format-fix": "prettier --write *",
     "lint-check": "eslint '**/*.{ts,tsx,js,jsx}'",
     "lint-fix": "eslint '**/*.{ts,tsx,js,jsx}' --fix",
-    "tests": "node ./test-custom-rules.js",
+    "tests": "node ./custom-rules/relative-extensions.test.js",
     "test": "npm run format-check && npm run lint-check && npm run tests"
   },
   "repository": {
